cordova-angular: use new pause event API
Uses the new 'pauseLock.release()' API to suspend the app only after
the socket.io server closes.
Keeps track of every socket to close the underlying connection in the
pause event.
Adds a error event listener to not crash when the OS forcibly closes
an unclosed socket when the application suspends.